Remove .png from repository
It is not good practice to track output collateral in a git project.
Every time the code is run, it generates new PNG files, and Git would
consider them changed. This means either you would have to constantly
revert the PNG to the original state, or constantly be commiting changed
PNG files that are only different in the metadata. The proper solution
is to simply not track them.

If one is bothered by them showing up as untracked, they can be added
to a .gitignore file, but in general I prefer to not keep a .gitignore
file.