Add requirements.txt file
A requirements.txt file is a collection of a project's dependencies in a
single file that can be passed to pip for installation. It makes it easy
to communicate and realize what is needed to run a project.

In this case, I have created a requirements.in which includes only the
libraries directly used by the project, and then used the command-line
tool "pip-compile" (from the pip-tools project) to create the
requirements.txt with ALL the libraries needed, including my
dependencies' dependencies. Each library is "pinned" with a specific
version so that results can be reproduced.