Add python equivalent functions for Mie theory calculations
The Fortran code has been completely ported to Python. To ensure
accuracy, all functions are tested using the hypothesis library to
generate random input, and the Python and Fortran implementations should
generate the same results.

As part of this, the Python implementations are compiled with a
"just-in-time" compilation library called numba. Basically, compilation
happens at runtime when the function is first called. The functions are
annotated with the "numba.jit" decorator, which at runtime will analyze
the function, deduce type information, and compile to native machine
code. The hope is that this causes the speed to approach that of
Fortran.